refactor: search screen 컴포넌트에서 최근 검색어 컴포넌트 분리 (#160)
