+++ apps/web/components/screen/SearchScreen/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchScreen } from './SearchScreen';
