+++ apps/web/components/screen/index.ts
@@ -1,1 +1,1 @@
-export { default as SearchScreen } from './SearchScreen';
+export { SearchScreen } from './SearchScreen';
